Fixed MySQL server has gone away issue
Using executemany might build a massive query on the mysql session if
the data to import is large. This will cause the entire query the buckle
and throw an OperationalError 'MySQL server has gone away'.

This approach might not be the most efficient, but does mean that the
import doesnt fail due to a large data import.